BO: Fix combination generator according to new combination system
